--- tunikov.dmitrii/lab3/node/Node.h
@@ -1,18 +1,0 @@
-//
-// Created by dmitrii on 15.12.2019.
-//
-
-#ifndef LAB3_NODE_H
-#define LAB3_NODE_H
-
-#include <memory>
-
-struct Node {
-    std::shared_ptr<int> m_data;
-    Node* m_next;
-    explicit Node(const int& data);
-    static void delete_nodes(Node* node);
-};
-
-
-#endif //LAB3_NODE_H
