+++ tunikov.dmitrii/lab3/stack/Stack.h
@@ -8,7 +8,7 @@
 #include <memory>
 #include <vector>
 
-enum StackType
+enum class StackType
 {
     LOCK,
     LOCK_FREE
@@ -21,8 +21,6 @@ public:
     virtual std::shared_ptr<int> pop() = 0;
     virtual void push(const int& val) = 0;
     virtual bool empty() = 0;
-    static void *readFromStack(void *arg);
-    static void* writeToStack(void *arg);
     virtual ~Stack();
 };
